requirements.txt
flask
nump
scikit-learn
gunicorn
Flask-Cors
pandas